requirements_api.txt
fastapi
uvicorn[standard]
python-multipart
numpy
pillow
torch
torchvision
huggingface_hub
transformers
